+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## 0.0.52
+- Input v-if for masking (attempt at length fix)
+- Load field request
 ## 0.0.51
 - Dependson updates and fixes
 ## 0.0.50